Change to object oriented programming
It is still a draft.

GameManager contains the matrix of the board.

It is now thought to be used to for a list of n players, each one can be either a user or a computer.

It is also thought to be used for different games: Forza Quattro, Tic Tac Toe. In other words: different number of lines or columns and also different number of marks to win.

The tris file is still in the project, but will be removed in the next updates.