Fix bugs and automatic users more realistic
Some print in main.py were shown when not necessary.

In game_manager.py, if the matrix was full, it printed "tie" even if there was a winner: the method is_tie() was just checking if the matrix was full. Now the name has been changed in is:matrix_full().

Automatic player will delay their answer of a few seconds, like a human user would do. In case of a future update that concerns a very intelligent player mode that takes a lot of execution time, the waiting time purposely added, is reduced by the duration of that execution.